WellKnownGeometry

Documentation for WellKnownGeometry.

WellKnownGeometry._getwkb!

WellKnownGeometry._getwkt!

WellKnownGeometry.getwkb

WellKnownGeometry.getwkb!

WellKnownGeometry.getwkt

WellKnownGeometry.getwkt!

WellKnownGeometry._getwkb! — Method

Push WKB to data for non Pointlike type of geom.

first indicates whether we need to indicate the type in case this outer geometry. repeat indicates whether sub geometries need to indicate their type, in case geom is a geometrycollection.

WellKnownGeometry._getwkt! — Method

Push WKT to data for non Pointlike type of geom.

first indicates whether we need to print the type with brackets–like POLYGON ( )– in case this outer geometry. repeat indicates whether sub geometries need to print their type, in case geom is a geometrycollection.

WellKnownGeometry.getwkb! — Method

getwkb!(data, type::T, geom, first::Bool)

Push WKB to data for a Pointlike type of geom.

first indicates whether we need to indicate the type in case this outer geometry or part of a GeometryCollection.

WellKnownGeometry.getwkb — Method

getwkb(geom)

Retrieve the Well Known Binary (WKB) as Vector{UInt8} for a geom that implements the GeoInterface.

WellKnownGeometry.getwkt! — Method

Push WKT to data for a Pointlike type of geom.

first indicates whether we need to print the type with brackets–like POINT ( )– in case this outer geometry or part of a geometrycollection.

WellKnownGeometry.getwkt — Method

getwkt(geom)

Retrieve the Well Known Text (WKT) as String for a geom that implements the GeoInterface.